fix flyers-form-demo.php
